chore: move benchmarks
