+++ packages/style-value-parser/__benchmarks__/calc-constant.bench.mjs
@@ -8,8 +8,8 @@
  */
 
 import Benchmark from 'benchmark';
-import { calcConstant as calcConstantLegacy } from '../css-types/calc-constant';
-import { calcConstant as calcConstantNew } from '../css-types-from-tokens/calc-constant';
+import { calcConstant as calcConstantLegacy } from '../lib/css-types/calc-constant.js';
+import { calcConstant as calcConstantNew } from '../lib/css-types-from-tokens/calc-constant.js';
 
 const calcConstantSuite = new Benchmark.Suite('CalcConstant');
