+++ packages/style-value-parser/__benchmarks__/dashed-ident.bench.mjs
@@ -8,8 +8,8 @@
  */
 
 import Benchmark from 'benchmark';
-import { DashedIdentifier as DashedIdentifierLegacy } from '../css-types/dashed-ident';
-import { DashedIdentifier as DashedIdentifierNew } from '../css-types-from-tokens/dashed-ident';
+import { DashedIdentifier as DashedIdentifierLegacy } from '../lib/css-types/dashed-ident.js';
+import { DashedIdentifier as DashedIdentifierNew } from '../lib/css-types-from-tokens/dashed-ident.js';
 
 const dashedIdentSuite = new Benchmark.Suite('DashedIdentifier');
