+++ packages/style-value-parser/__benchmarks__/dimension.bench.mjs
@@ -8,8 +8,8 @@
  */
 
 import Benchmark from 'benchmark';
-import { dimension as dimensionLegacy } from '../css-types/dimension';
-import { dimension as dimensionNew } from '../css-types-from-tokens/dimension';
+import { dimension as dimensionLegacy } from '../lib/css-types/dimension.js';
+import { dimension as dimensionNew } from '../lib/css-types-from-tokens/dimension.js';
 
 const dimensionSuite = new Benchmark.Suite('Dimension');
