+++ packages/style-value-parser/__benchmarks__/length-percentage.bench.mjs
@@ -8,8 +8,8 @@
  */
 
 import Benchmark from 'benchmark';
-import { lengthPercentage as lengthPercentageLegacy } from '../css-types/length-percentage';
-import { lengthPercentage as lengthPercentageNew } from '../css-types-from-tokens/length-percentage';
+import { lengthPercentage as lengthPercentageLegacy } from '../lib/css-types/length-percentage.js';
+import { lengthPercentage as lengthPercentageNew } from '../lib/css-types-from-tokens/length-percentage.js';
 
 const lengthPercentageSuite = new Benchmark.Suite('LengthPercentage');
